Refactor exchange limits
The trouble with the original map based approach was that it didn't
provide a good way of specifying currency agnostic defaults for
exchanges that don't specify the limits in their API docs.

As a bonus there's no longer any need to format a currency pair just
right in the calling code in order to retrieve the limits for a
currency particular pair.