Convert Bittrex symbols to normalized currency pairs and back
The market symbols Bittrex uses look like currency pairs but don't
follow the common conventions for currency pairs. For example on
most exchanges the ETH/BTC pair means the trade amount is denoted in
ETH, and the trade price is denoted in BTC (regardless of whether
you're buying or selling). However, on Bittrex to trade ETH you have
to use the BTC-ETH market, yet the trade amount is denoted in ETH,
and the trade price is denoted in BTC, which is the exact opposite
of the common conventions.

Users of this library shouldn't have jump through hoops to deal with
Bittrex so currency pairs returned and accepted via the IBotExchange
and IBotExchangeEx interfaces now have the same meaning as on other
exchanges.