Implement Bitfinex.CalcAvailableBalance()
This is primarily useful for margin orders since it makes it easy
to figure the max amount for the margin order. Sadly Bitfinex seems
to error out when you attempt to use the same API key for both
v1 and v2 API endpoints, so it's currently rather difficult to
actually use this new functionality since all the other API
endpoints are still v1.