Add pairs parameter to IBotExchangeEx.GetOrders()
This change was made to try and avoid hitting Binance rate limiting
because they consider fetching open orders for all symbols to be
an expensive operation. Some of the other exchanges also support
retrieving a subset of open orders based on the symbol, but that
will need to be implemented later.