Check HTTP status code to detect Bitfinex rate-limiting
While the API docs seem to indicate Bitfinex will send a rate-limit
error as JSON it's just as likely to do so via Cloudflare - in which
case you get an HTML page sent back and a 429 HTTP status code.
These changes detect the 429 status code so that the rate-limiting
code can block any further requests for a bit.